pyproject.toml
[tool.poetry]
name = "telegram-work-bot"
version = "0.1.0"
description = ""
authors = ["Sorokin Vadym <s57111702@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyTelegramBotAPI = "^4.5.1"
feedparser = "^6.0.10"
SQLAlchemy = "^1.4.36"
aiohttp = "^3.8.1"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
